% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesReact_core.R
\name{bayesReact_core}
\alias{bayesReact_core}
\title{bayesReact method for motif activity inference}
\usage{
bayesReact_core(
  lst_data,
  threshold_motif_prob = 1e-10,
  threshold_motif_count = 2,
  model = "bayesReact",
  output_type = "activity_summary",
  CI = c(0.1, 0.9),
  MCMC_iterations = 3000,
  MCMC_chains = 3,
  MCMC_warmup = 500,
  MCMC_cores = MCMC_chains,
  MCMC_keep_warmup = F,
  parallel = F
)
}
\arguments{
\item{lst_data}{list containing file paths or data frames, e.g.: list(FC_rank = "./FC_rank_date.rds", motif_probs = "./seqXmot_probs.rds", motif_counts = "./seqXmot_counts.rds").}

\item{threshold_motif_prob}{numeric value specifying the minimum threshold used to truncate the probability of observing a motif at least once in a sequence. If set to NULL, no threshold is used.}

\item{threshold_motif_count}{integer specifying the maximum threshold used to truncate the number of times a motif is observed in a sequence. If set to NULL, no threshold is used.}

\item{model}{A character string specifying the model to be used for inference:
Either "bayesReact" (default) or "BF" (Bayes Factor; when comparing beta model against the uniform null model).}

\item{output_type}{type of output to be returned; either "activity" (only outputs activity score); "activity_summary" (default; outputs activity score, posterior mean and sd for the underlying activity parameter 'a', log P(|a| <= 0 | data), credible intervals, and simple model diagnostics);
"full_posterior" (outputs all MCMC iterations after warm-up period); or "full_model" (returns full stanfit model object). "full_posterior" and "full_model" can only be obtained for a single motif at a time.}

\item{CI}{credible interval (CI) to be returned for activity estimates. Default is 80\% CI: c(0.10, 0.90).}

\item{MCMC_iterations}{number of iterations to be run by the MCMC sampler.}

\item{MCMC_chains}{number of independent MCMC chains to be used.}

\item{MCMC_warmup}{initial iterations to be discarded for each chain as warm-up/burn-in.}

\item{MCMC_cores}{number of cores, default is equal to the number of chains (which is the maximum number of cores that can be utilized by STAN's MCMC sampler).
Alternatively consider parallel::detectCores().}

\item{MCMC_keep_warmup}{whether to keep the warm-up iterations or not.}

\item{parallel}{this parameter should never be changed manually and is used internally by bayesReact_parallel().}
}
\value{
Motif activity estimates in a format specified by output_type.
}
\description{
Function for predicting motif activity from ranked sequence data using a simple Bayesian model implemented in STAN.
bayesReact_core can be run locally and used for motif activity inference for a set of motifs across a small data set.
For larger data sets, please consider bayesReact_parallel().
}
\examples{
\dontrun{
motif_activities <- bayesReact_core(lst_data =
list(FC_rank = FC_rank, motif_probs = motif_probs, motif_counts = motif_counts))
}

}
