% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_motif_model.R
\name{construct_motif_model}
\alias{construct_motif_model}
\title{Wrapper function to construct a STAN model for motif activity inference.}
\usage{
construct_motif_model(model = "bayesReact")
}
\arguments{
\item{model}{A character string specifying the model to be used for inference:
Either "bayesReact" (default) or "BF" (Bayes Factor; when comparing support for the beta model against the uniform null model).}
}
\value{
A stanmodel object.
}
\description{
This function constructs a STAN model object for motif activity inference, and is used by fit_motif_model(), bayesReact_core(), and bayesReact_parallel().
}
\examples{
\dontrun{
my_model <- construct_motif_model()
}

}
