% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_kmer_counting.R
\name{count_all_kmers_nonoverlap}
\alias{count_all_kmers_nonoverlap}
\title{Fast exact non-overlapping counting of all k-mers using C++.}
\usage{
count_all_kmers_nonoverlap(k, seqlist)
}
\description{
Calls ./src/fast_kmers_counting.cpp
}
\keyword{internal}
