% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_motif_model.R
\name{fit_motif_model}
\alias{fit_motif_model}
\title{Fit model for motif activity prediction}
\usage{
fit_motif_model(
  input,
  model,
  model_type = "bayesReact",
  output_type = "activity_summary",
  CI = c(0.1, 0.9),
  iterations = 3000,
  chains = 3,
  warmup = 500,
  cores = chains,
  keep_warmup = F
)
}
\arguments{
\item{input}{path to input file (.rds format) or list object produced by bayesReact::prep_model_input().}

\item{model}{STAN model object to be used for motif activity inference. The stanmodel object is constructed by construct_motif_model().}

\item{model_type}{a character string specifying the type of stanmodel object constructed by construct_motif_model():
Either "bayesReact" (default) or "BF" (Bayes Factor; when comparing beta model against uniform null model).}

\item{output_type}{type of output to be returned; either "activity" (only outputs activity score); "activity_summary" (default; outputs activity score, posterior mean and sd for the underlying activity parameter 'a', log P(|a| <= 0 | data), credible intervals, and simple model diagnostics);
"full_posterior" (outputs all MCMC iterations after warm-up period); or "full_model" (returns full stanfit model object).}

\item{CI}{credible interval (CI) to be returned for activity estimates. Default is 80\% CI: c(0.10, 0.90).}

\item{iterations}{number of iterations to be run by the MCMC sampler.}

\item{chains}{number of independent MCMC chains to be run.}

\item{warmup}{initial iterations to be discarded for each chain as warm-up/burn-in.}

\item{cores}{number of cores, default is equal to the number of chains (which is the maximum number of cores that can be utilized by STAN's MCMC sampler). Alternatively, consider parallel::detectCores().}

\item{keep_warmup}{whether to keep the warm-up iterations or not.}
}
\value{
Activity estimates in a format specified by output_type.
}
\description{
Function for predicting motif activity from ranked sequence data using a simple Bayesian model implemented in STAN.
}
\examples{
\dontrun{
my_activity <- fit_motif_model(my_data)
# returns df with activity estimates, CIs, and simple diagnostics
}

}
