% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miReact_code.R
\name{miReact_code}
\alias{miReact_code}
\alias{pd.mrs2}
\alias{test2}
\title{miReact code dependencies}
\usage{
miReact_code()

pd.mrs2(pattern, seq)

test2(pattern, seq)
}
\arguments{
\item{pattern}{state space (transition matrix) for motif of interest.}

\item{seq}{sequence for which the probability of motif occurrence is evaluated.}
}
\value{
motif probability (pd.mrs2)
or X (test2).
}
\description{
This constitutes some of the core miReact code obtained from https://github.com/muhligs/miReact (19/12/2023).
For more information on miReact, please see their publication: https://doi.org/10.1038/s41598-021-88480-5 (Nielsen et al., Sci Rep, 2021).
bayesReact depends on both Regmex and miReact for handling input data and parallelization using the Slurm queuing system when running on a computer clusters.
}
\section{Functions}{
\itemize{
\item \code{pd.mrs2()}: Obtain motif probability in a given sequence.

\item \code{test2()}: Test.

}}
\examples{
# See 'https://github.com/muhligs/miReact'.

}
