% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_model_input.R
\name{prep_model_input}
\alias{prep_model_input}
\title{Helper function to prepare input data used for model fitting and motif activity estimation}
\usage{
prep_model_input(
  in_seq_motif_data,
  threshold_motif_prob = 1e-10,
  threshold_motif_count = 2
)
}
\arguments{
\item{in_seq_motif_data}{list containing FC_rank (seq X sample), motif_probs (1 X seq), motif_counts (1 X seq).}

\item{threshold_motif_prob}{numeric value specifying the minimum threshold used to truncate the probability of observing a motif at least once in a sequence. If set to NULL, no threshold is used.}

\item{threshold_motif_count}{integer specifying the maximum threshold used to truncate the number of times a motif is observed in a sequence. If set to NULL, no threshold is used.}
}
\value{
List containing all necessary variables used as input for the STAN model.
}
\description{
This function generate the appropriate input for the STAN model used to estimate motif activity.
}
\examples{
\dontrun{
inlist <- prep_model_input(in_seq_motif_data =
list(FC_rank = FC_rank, motif_probs = motif_probs, motif_counts = motif_counts))
}

}
