% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_input.R
\name{process_raw_input}
\alias{process_raw_input}
\title{Function to process raw input data needed to run bayesReact}
\usage{
process_raw_input(
  exp,
  seq_data,
  motifs,
  out_path = "./",
  exp_type = "count",
  save_processed_exp = F,
  process_all = T,
  seq_gene_id = "gsym",
  seq_min_length = 20,
  seq_max_length = 10000,
  approx_motif_prob = F,
  cores = parallel::detectCores()
)
}
\arguments{
\item{exp}{numerical gene expression (or other relevant measure) matrix (rows = genes, columns = samples) or character string specifying file path to expression data (saved in .rds format). Input should NOT be log-transformed. Examples include raw counts (default), CPM, or TPM values.}

\item{seq_data}{character string, data frame or vector; path to a bioMart output-file (decompressed .fasta file); a data frame containing gene_name/gene_ID, sequences and sequence length;
or a character vector of length two, containing the paths to seqs and seqlist objects (e.g., seq_data = c("./seqs.rds", "./seqlist.rds")).}

\item{motifs}{vector of motifs (e.g., c("ACGTAGT", "GTACAAG")) or an integer specifying the k-mers to be used (e.g., 7 for all 7-mers).
A motif is specified as a regular expression on the alphabet {A,C,G,T}.}

\item{out_path}{character string specifying directory to save .rds files in. Default is current working directory.}

\item{exp_type}{character string specifying type of data to normalize and/or scale. Input can be "count" (default) or "CPM" (counts per million; TPM can also be used).}

\item{save_processed_exp}{logical, whether to also save the normalized and scaled matrix. Default is FALSE.}

\item{process_all}{logical, whether to process all data (expression, sequence, and motif data) or only computing the sample-specific fold-change ranks.}

\item{seq_gene_id}{character specifying the gene ID that should be used to build the sequence list. The IDs has to match the IDs in the gene/transcript expression matrix.
The input can be "gsym" (gene symbol, default), "gid" (ensembl gene id), "tid" (ensembl transcript id). Only relevant to specify when using a bioMart file-path as input.}

\item{seq_min_length}{integer, minimum sequence length to include in the seqs object and list. Default is 20.}

\item{seq_max_length}{integer, maximum sequence length to include in the seqs object and list. Default is 10,000.}

\item{approx_motif_prob}{logical, whether to use a binomial approximation to compute motif probabilities (default is FALSE).}

\item{cores}{integer, number of cores to use for parallel processing when computing motif probabilities (default is all available cores).}
}
\value{
process_raw_input() calculates and saves the Fold-change ranks, processed sequence data and sequence-specific motif probabilities, and returns a list containing all file paths; list(FC_rank_path, seqs_path, seqlist_path, motif_prob_path).
}
\description{
The function processes raw expression, motif and sequence data, in order to produce sequence ranks and motif probabilities needed for numerical sequence representation.
The processed data is used by bayesReact to evaluate motif distribution across the expression (or other continuous measure) ranked sequences.
The function calls the following sub-functions: norm_scale_seq, rank_seq, build_seq_list, and motif_prob, which can also be run individually.
}
\examples{
\dontrun{
out_file_paths <- process_raw_input(exp = "./project/folder/exp_counts.rds", seq_data = seqs,
motifs = 7, out_path = "./project/folder/")
}

}
